+++ server/pkg/oauth/api/api_suite_test.go
@@ -1,0 +1,13 @@
+package api_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestAuthenticate(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Oauth Api Test Suite")
+}
